Move typings from node to common
