+++ packages/cli-kit/src/metadata.ts
@@ -1,7 +1,7 @@
 import {AnyJson} from './private/common/json.js'
 import {MonorailEventPublic} from './monorail.js'
 import {sendErrorToBugsnag} from './public/node/error-handler.js'
-import {PickByPrefix} from './public/node/typing/pick-by-prefix.js'
+import {PickByPrefix} from './public/common/typing/pick-by-prefix.js.js'
 import {isUnitTest} from './public/node/environment/local.js'
 
 type ProvideMetadata<T> = () => Partial<T> | Promise<Partial<T>>
